Add explicit type conversions where necessary: media/
These are cases that are implicitly narrowing today, and must do so
explicitly in order to enable -Wc++11-narrowing.  No behavior change
intended.

Bug: 1216696
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2948073
Auto-Submit: Peter Kasting <pkasting@chromium.org>
Commit-Queue: Will Cassella <cassew@google.com>
Cr-Commit-Position: refs/heads/master@{#890792}

+++ media/capture/video/win/gpu_memory_buffer_tracker.cc
@@ -22,8 +22,8 @@ base::win::ScopedHandle CreateNV12Texture(ID3D11Device* d3d11_device,
                                           const gfx::Size& size) {
   const DXGI_FORMAT dxgi_format = DXGI_FORMAT_NV12;
   D3D11_TEXTURE2D_DESC desc = {
-      .Width = size.width(),
-      .Height = size.height(),
+      .Width = static_cast<UINT>(size.width()),
+      .Height = static_cast<UINT>(size.height()),
       .MipLevels = 1,
       .ArraySize = 1,
       .Format = dxgi_format,
